Started on native screenshots feature, added wip doodle screenshot saver which currently can save screens of all graphics chips but c64dtv-vicii and vdc
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@22579 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/plus4/tedtypes.h
@@ -193,7 +193,7 @@ struct ted_s {
     raster_t raster;
 
     /* VIC-II registers.  */
-    int regs[64];
+    BYTE regs[64];
 
     /* Interrupt register.  */
     int irq_status;             /* = 0; */
